add cloudinary to upload images and allow admin to upload new pizzas

+++ pages/api/products/[id].js
@@ -5,7 +5,10 @@ export default async function handler(req, res) {
   const {
     method,
     query: { id },
+    cookies,
   } = req;
+
+  const token = cookies.token;
 
   // connect db
   dbConnect();
@@ -18,6 +21,10 @@ export default async function handler(req, res) {
       res.status(500).json(error);
     }
   } else if (method === 'PUT') {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json('Not Authenticated!');
+    }
+
     try {
       const product = await Product.findByIdAndUpdate(id, req.body, {
         // to return the new updated order
@@ -28,6 +35,10 @@ export default async function handler(req, res) {
       res.status(500).json(error);
     }
   } else if (method === 'DELETE') {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json('Not Authenticated!');
+    }
+
     try {
       await Product.findByIdAndDelete(id);
       res.status(200).json('The product has been deleted!');
